+++ index.css
@@ -189,8 +189,7 @@ h3 {
   float: right;
 }
 
-.date,
-.time {
+.date {
   color: green;
 }
