+++ index.html
@@ -22,19 +22,12 @@
     </header>
     <div id="dateTime">
         <span class="date"></span>
-        <span class="time"></span>
       </div>
     <section id="book-library">
         <h1 class="description">All awesome books</h1>
         <div class="main-container">
             <div id="book-collection">
-                <thead>
-                    <tr>
-                        <th></th>
-                        <th></th>
-                        <th></th>
-                    </tr>
-                </thead>
+                
                 <tbody id="booklist"></tbody>            
             </div>
         </div>
