+++ index.js
@@ -1,16 +1,7 @@
 import './modules/navigation.js';
-/* eslint-disable */
-import {
-  DateTime, Duration, FixedOffsetZone,
-   IANAZone, Info, Interval, InvalidZone,
-    Settings, SystemZone, VERSION, Zone,
-} from './modules/luxon.js';
-/* eslint-enable */
-import './modules/dateTime.js';
+import date from './modules/dateTime.js';
 import './modules/app.js';
+import './modules/script.js';
+import './modules/localStorage.js';
 
-// Validating form
-
-// ====== Handle local Storage =====
-
-// ==== Handle UI ====
+date();
